pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "SEO suitability webpages analyzer"
authors = ["Alina N <silumain@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Bishamontess/python-project-83"
classifiers = ["Private :: Webapp"]
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
